requirements.txt
streamlit
pytesseract
pillow
pdf2image
PyMuPDF
googletrans==4.0.0-rc1
opencv-python
numpy
